monad-discovery: extend name record with ports and capabilities
this change replaces single socket addr with the following fields:
- ip
- ports
- capabilities

ports is a collection, limited to 8 items, with port for tcp, udp.

capabilities is meant to be used for rolling upgrades. so for example
with direct socket, we will be able to use plain protocol selectively
and not necessarily at the same time when socket is enabled.

the other use case that i foresee in the future is the support of
authenticated protocol for udp connection, that would also benefit
from negotiated capability

this change doesn't switch to v2 immediately, but makes it possible to
decode it. in decode method of the NameRecord we try to decode v2 first
and if it fails we will try v1. NameRecord::new is using v1, in the
follow up change it will be switched to v2. And then in the next follow
up v1 will be scraped from the codebase.